Video Core: Add ignore format reinterpret hack
This is an intentional hack which will ignore reloading textures
from the cpu when the texture was created solely on the gpu (ie from a
texture copy) and the requested texture format differs from cached
format. Normally this would require a flush as the formats are different
but this will skip that bringing huge performance benefits to games
that use this to draw mostly unnoticable after effects.

Temporary! Log pixel format when ignoring format reinterpretation

Used to debug cases when hw texture encoding / decoding will be useful
and when it won't work (such as if the depth buffer is being read from
in a texture format)

Add telemetry to see when the hack is used

This just sounds like nonsense to me. I'm sure devs aren't sitting
around going "hmm now what?". This sounds like a solution in search of a
problem.