feat(bps): Backport BPS support (#2)
* file_sys: Move IPS patching code into separate source file

In anticipation of a new BPS patcher.

* file_sys: Handle patch applying failures

This changes ApplyCodePatch to return a ResultStatus, which makes it
possible to determine whether patch applying has failed. Previously,
only a boolean was returned, and false was returned when no patch
was found OR when a patch was found but applying it failed.

This also changes AppLoader_NCCH to return an error if patching fails
because the executable is likely to be left in an inconsistent state
and we should not proceed booting in that case.

* file_sys: Add support for BPS patches

The BPS format allows distributing patches that are smaller and that do
not contain copyrighted content if data is relocated
(unlike non-trivial IPS patches).

This is essential for games such as MM3D that have three barely
different code revisions. Supporting all three versions would
demand an unreasonable amount of work; with BPS patches only one
version has to be supported.

* Resolve some type conversion warnings

* fix(bps): Fixes BPS patch changing target size (#5829)

Co-authored-by: Léo Lam <leo@leolam.fr>
Co-authored-by: Vitor Kiguchi <vitor-kiguchi@hotmail.com>