feat: Backport LayeredFS (#8)
* core/file_sys: LayeredFS implementation

This implementation is different from Luma3DS's which directly hooks the SDK functions. Instead, we read the RomFS's metadata and figure out the directory and file structure. Then, relocations (i.e. replacements/deletions/patches) are applied. Afterwards, we rebuild the metadata, and assign 'fake' data offsets to the files. When we want to read file data from this rebuilt RomFS, we use binary search to find the last data offset smaller or equal to the given offset and read from that file  (either from the original RomFS, or from replacement files, or from buffered data with patches applied) and any later files when length is not enough.

The code that rebuilds the metadata is pretty complex and uses quite a few variables to keep track of necessary information like metadata offsets. According to my tests, it is able to build RomFS-es identical to the original (but without trailing garbage data) when no relocations are applied.

* core: Use LayeredFS while reading RomFS

Only enabled for NCCHs that do not have an override romfs.

LayeredFS files should be put in the `load` directory in User Directory. The directory structure is similar to yuzu's but currently does not allow named mods yet. Replacement files should be put in `load/mods/<Title ID>/romfs` while patches/stubs should be put in `load/mods/<Title ID>/romfs_ext`.

* citra_qt: Add 'Open Mods Location'

* core/file_sys: Read mods for the original title for updates

Updates can override RomFS and ExeFS, therefore we should apply the mods to them as well.

* Fix build

Explicitly use `std::min<std::size_t>`

Added virtual destructor

* core: Reset archive_manager on shutdown.

This holds the archives which include the SelfNCCH archive which holds the RomFS files. If we don't reset it the LayeredFS class can't get destructed and mods files won't be released.

* Address review

simplify code

* core: Add dump RomFS support

This is added to LayeredFS, then the NCCH container and then the loader interface.

* citra_qt: Add 'Dump RomFS' menu action

A progress dialog will be displayed. However no progress is reported and the user also cannot cancel it.

* layered_fs: Do not open all replacement files on load

Instead open them when we want to read them. This is because the standard library has a limit on the number of opened files.

* fix clang format

* core/file_sys: Do not apply the same mods to DLCs

Now you can apply separate mods to DLCs and mods for the original title won't be applied.

* layered_fs: Fix missing file size update

This was a silly typo from a previous change.

* fix: Fixes incorrect RomFS reader reference

Co-authored-by: zhupengfei <zhupf321@gmail.com>